Rename the Q# test loader and drop the stale staging wording
qsharp_test_project.py was named for the temp Q# project it used to build; that
project is gone, so the name described nothing. qsharp_test_sources.py pairs
with the TEST_SOURCE_ROOT it exports, and unlike test_qsharp_project.py it stays
out of pytest's default test_*.py collection - it holds no tests.

The doc comments on the three .qs files, the _test_utils docstring and
test_the_staged_test_context_can_reach_them still claimed the drivers were
staged into a throwaway project. They are evaluated onto a throwaway context
instead. The reason for doing it that way is unchanged and still recorded on
UnaryPhaseEstimationTests.qs: internal visibility is package-scoped, so a
dependent package could not reach ApplySignedPowerSchedule.
